+++ phoenix-core/src/main/java/org/apache/phoenix/schema/PTable.java
@@ -683,7 +683,7 @@ public interface PTable extends PMetaDataEntity {
 
     ViewType getViewType();
     String getViewStatement();
-    Short getViewIndexId();
+    Long getViewIndexId();
     PTableKey getKey();
 
     IndexType getIndexType();
